requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
djangorestframework==3.15.2
djangorestframework-jsonapi==7.1.0
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
packaging==24.2
pi==0.1.2
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3